requirements.txt
# Core AI/ML
langchain==0.3.21
langchain-core==0.3.58
langchain-community==0.3.20
openai==1.70.0
google-generativeai==0.8.4
cohere==5.14.0

# Vector Database & ML Runtime
chromadb==0.6.3
pysqlite3-binary>=0.5.1
onnxruntime==1.17.0

# Document Processing
PyMuPDF==1.25.4
python-dotenv==1.1.0

# Web Interface
streamlit==1.43.2

# Utils
pydantic==2.11.2
numpy==1.26.4
pandas==2.2.3